fix(Platform, Tests, Provider): Deprecated Platform.remove() in favor of Platform.delete(); Changed return value of Platform methods to be inline with what the method does, now the methods returned the updated value. Added tests for Platform methods and Setup flags. Added Setup flag to provider, to avoid multiple calls to setup method and avoid deployment without setup. Added deploy flag to Database, to avoid usage without deployment

+++ src/Utils/Database.js
@@ -9,6 +9,7 @@ const provDatabaseDebug = require('debug')('provider:database')
  */
 class Database {
   #dbConnection = {}
+  #deploy = false
 
   /**
    * @description Mongodb configuration setup
@@ -144,6 +145,7 @@ class Database {
     })
 
     if (this.db.readyState === 0) await mongoose.connect(this.#dbConnection.url, this.#dbConnection.options)
+    this.#deploy = true
     return true
   }
 
@@ -151,6 +153,7 @@ class Database {
   async Close () {
     mongoose.connection.removeAllListeners()
     await mongoose.connection.close()
+    this.#deploy = false
     return true
   }
 
@@ -161,6 +164,7 @@ class Database {
      * @param {Object} [query] - Query for the item you are looking for in the format {type: "type1"}.
      */
   async Get (ENCRYPTIONKEY, collection, query) {
+    if (!this.#deploy) throw new Error('PROVIDER_NOT_DEPLOYED')
     if (!collection) throw new Error('MISSING_COLLECTION')
 
     const Model = mongoose.model(collection)
@@ -189,6 +193,7 @@ class Database {
      * @param {Object} [index] - Key that should be used as index in case of Encrypted document.
      */
   async Insert (ENCRYPTIONKEY, collection, item, index) {
+    if (!this.#deploy) throw new Error('PROVIDER_NOT_DEPLOYED')
     if (!collection || !item || (ENCRYPTIONKEY && !index)) throw new Error('MISSING_PARAMS')
 
     const Model = mongoose.model(collection)
@@ -215,6 +220,7 @@ class Database {
    * @param {Object} [index] - Key that should be used as index in case of Encrypted document.
    */
   async Replace (ENCRYPTIONKEY, collection, query, item, index) {
+    if (!this.#deploy) throw new Error('PROVIDER_NOT_DEPLOYED')
     if (!collection || !item || (ENCRYPTIONKEY && !index)) throw new Error('MISSING_PARAMS')
 
     const Model = mongoose.model(collection)
@@ -240,6 +246,7 @@ class Database {
      * @param {Object} modification - The modification you want to make in the format {type: "type2"}.
      */
   async Modify (ENCRYPTIONKEY, collection, query, modification) {
+    if (!this.#deploy) throw new Error('PROVIDER_NOT_DEPLOYED')
     if (!collection || !query || !modification) throw new Error('MISSING_PARAMS')
 
     const Model = mongoose.model(collection)
@@ -264,6 +271,7 @@ class Database {
      * @param {Object} query - The entry you want to delete in the format {type: "type1"}.
      */
   async Delete (collection, query) {
+    if (!this.#deploy) throw new Error('PROVIDER_NOT_DEPLOYED')
     if (!collection || !query) throw new Error('MISSING_PARAMS')
     const Model = mongoose.model(collection)
     await Model.deleteMany(query)
